requirements.txt
# Essential dependencies for document processing
google-cloud-documentai==2.21.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit==1.28.0